ref #28890 - throw Exception in DepartmentDelete Api Controller if entity has children

+++ src/Zmsapi/DepartmentDelete.php
@@ -1,30 +1,39 @@
 <?php
-/**
- * @package 115Mandant
- * @copyright BerlinOnline Stadtportal GmbH & Co. KG
- **/
 
+/**
+ *
+ * @package Zmsadmin
+ * @copyright BerlinOnline Stadtportal GmbH & Co. KG
+ *
+ */
 namespace BO\Zmsapi;
 
 use \BO\Slim\Render;
-use \BO\Mellon\Validator;
 use \BO\Zmsdb\Department as Query;
 
 /**
-  * Handle requests concerning services
-  */
+ * Delete a department
+ */
 class DepartmentDelete extends BaseController
 {
+
     /**
+     *
      * @return String
      */
     public static function render($itemId)
     {
         $query = new Query();
         $message = Response\Message::create(Render::$request);
-        $department = $query->readEntity($itemId);
-        $query->deleteEntity($itemId);
-        $message->data = $department;
+        $entity = $query->readEntity($itemId);
+        if ($entity->toProperty()
+            ->offsetExists('scopes') || $entity->toProperty()
+            ->offsetExists('clusters')) {
+            throw new Exception\Department\ScopeListNotEmpty();
+        } else {
+            $query->deleteEntity($itemId);
+        }
+        $message->data = $entity;
         Render::lastModified(time(), '0');
         Render::json($message->setUpdatedMetaData(), $message->getStatuscode());
     }
